Add files via upload
Base game.
I hate my code. I'd like to start over with better logic and structure and minimize redundant code and separate to more small function with understandable names.
As for the CSS/HTML. I simply gave up. I found it too difficult to make a decent page. And I wasn't happy about the fact that I had to spend so much time on HTML/CSS just to start over with something simple. My CSS is also a complete mess and lots of unneeded parameters.

I'd like to start over and do the bonuses as well over the weekend.